vulnseek/requirements.txt
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
configparser==5.3.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
psycopg2-binary==2.9.6
requests==2.31.0
tomli==2.0.1
urllib3==2.0.3
